with dir restructure
