+++ packages/text-area/src/TextArea/index.ts
@@ -1,0 +1,2 @@
+export { TextArea as default } from './TextArea'
+export { State } from './TextArea.types';
