+++ packages/text-area/src/index.ts
@@ -1,2 +1,2 @@
-export { TextArea as default } from './TextArea';
-export { State } from './types';
+export { default } from './TextArea';
+export { State } from './TextArea';
